Update yarn.lock via yarn update
The yarn.lock file had gone out of sync with the package.json file. I
was getting errors like the following when trying to run the Yarn build.

```
Couldn't find any versions for "coc.nvim" that matches "^0.0.32" in our
cache (possible versions are ""). This is usually caused by a missing
entry in the lockfile, running Yarn without the --offline flag may help
fix this issue.
```